+++ src/backend/cuda/CudaConfig_gen.h
@@ -118,6 +118,7 @@ size_t inline generate<Algorithm::RANDOM_X>(Threads<CudaThreads> &threads, const
     auto wow = CudaThreads(devices, Algorithm::RX_WOW);
     auto arq = CudaThreads(devices, Algorithm::RX_ARQ);
     auto kva = CudaThreads(devices, Algorithm::RX_KEVA);
+    auto xeq = CudaThreads(devices, Algorithm::RX_XEQ);
 
     if (!threads.isExist(Algorithm::RX_WOW) && wow != rx) {
         count += threads.move(Algorithm::kRX_WOW, std::move(wow));
@@ -129,6 +130,10 @@ size_t inline generate<Algorithm::RANDOM_X>(Threads<CudaThreads> &threads, const
 
     if (!threads.isExist(Algorithm::RX_KEVA) && kva != rx) {
         count += threads.move(Algorithm::kRX_KEVA, std::move(kva));
+    }
+
+    if (!threads.isExist(Algorithm::RX_XEQ) && xeq != rx) {
+        count += threads.move(Algorithm::kRX_XEQ, std::move(xeq));
     }
 
     count += threads.move(Algorithm::kRX, std::move(rx));