+++ src/base/crypto/Coin.h
@@ -44,6 +44,7 @@ public:
         ZEPHYR,
         TOWNFORGE,
         TUSKE,
+        XEQ,
         MAX
     };